@@ -1,179 +1,182 @@
-﻿namespace CodingTrackerConsole
+﻿namespace CodingTrackerConsole;
+
+public static class ProgramController
 {
-    public static class ProgramController
+    private static readonly DatabaseManager DbManager = new();
+    private static readonly UserInput Input = new();
+    private static readonly TableVisualizationEngine DisplayRecords = new();
+    private static readonly List<CodingTrackerModel> CurrentData = DataAccessService.LoadData(); // ************ //
+    private static readonly FilterController FilterController = new();
+
+    private static void DisplayMenu()
     {
-        private static readonly DatabaseManager DbManager = new();
-        private static readonly UserInput Input = new();
-        private static readonly TableVisualizationEngine DisplayRecords = new();
+        Console.WriteLine("MAIN MENU");
+        Console.WriteLine("---------------------------\n");
 
-        private static void DisplayMenu()
+        Console.WriteLine("What would you like to do?");
+        Console.WriteLine("c to Close Application");
+        Console.WriteLine("v to View All Records");
+        Console.WriteLine("i to Insert a Record");
+        Console.WriteLine("d to Delete a Record");
+        Console.WriteLine("u to Update a Record\n");
+
+        Console.WriteLine("Enter choice and hit Enter");
+        Console.Write("Your choice? ");
+    }
+
+    public static void StartProgram()
+    {
+        DbManager.CreateDatabase();
+
+        DisplayMenu();
+        string choice = Input.GetChoice();
+
+        while (choice != "c")
         {
-            Console.WriteLine("MAIN MENU");
-            Console.WriteLine("---------------------------\n");
-
-            Console.WriteLine("What would you like to do?");
-            Console.WriteLine("c to Close Application");
-            Console.WriteLine("v to View All Records");
-            Console.WriteLine("i to Insert a Record");
-            Console.WriteLine("d to Delete a Record");
-            Console.WriteLine("u to Update a Record\n");
-
-            Console.WriteLine("Enter choice and hit Enter");
-            Console.Write("Your choice? ");
-        }
-
-        public static void StartProgram()
-        {
-            DbManager.CreateDatabase();
+            switch (choice)
+            {
+                case "v":
+                    ViewRecords();
+                    break;
+                case "i":
+                    GetRecordsToInsert();
+                    break;
+                case "d":
+                    GetRecordsToDelete();
+                    break;
+                case "u":
+                    SelectRecordToUpdate();
+                    break;
+                default:
+                    Console.WriteLine("Wrong input!");
+                    break;
+            }
 
             DisplayMenu();
-            string choice = Input.GetChoice();
+            choice = Input.GetChoice();
+        }
+    }
 
-            while (choice != "c")
+    private static void GetRecordsToInsert()
+    {
+        Console.Clear();
+
+        var modelRecord = new CodingTrackerModel
+        {
+            Date = Input.GetDate(),
+            StartTime = Input.GetTime(),
+            EndTime = Input.GetTime(),
+        };
+        modelRecord.Duration = Input.GetDuration(modelRecord.StartTime, modelRecord.EndTime);
+
+        DbManager.InsertRecord(modelRecord);
+
+        ViewRecords();
+    }
+
+    private static void DisplayDeleteContextMenu()
+    {
+        Console.WriteLine("b to Go Back");
+        Console.WriteLine("d to Delete Record: ");
+    }
+
+    private static void GetRecordsToDelete()
+    {
+        Console.Clear();
+        ViewRecords();
+
+        DisplayDeleteContextMenu();
+        string choice = Input.GetChoice();
+        while (choice != "b")
+        {
+            switch (choice)
             {
-                switch (choice)
-                {
-                    case "v":
-                        ViewRecords();
-                        break;
-                    case "i":
-                        GetRecordsToInsert();
-                        break;
-                    case "d":
-                        GetRecordsToDelete();
-                        break;
-                    case "u":
-                        SelectRecordToUpdate();
-                        break;
-                    default:
-                        Console.WriteLine("Wrong input!");
-                        break;
-                }
-
-                DisplayMenu();
-                choice = Input.GetChoice();
+                case "d":
+                    DbManager.DeleteRecord(Input.GetDate());
+                    ViewRecords();
+                    break;
+                default:
+                    Console.WriteLine("Invalid Choice!");
+                    ViewRecords();
+                    break;
             }
-        }
-
-        private static void GetRecordsToInsert()
-        {
-            Console.Clear();
-
-            var modelRecord = new CodingTrackerModel
-            {
-                Date = Input.GetDate(),
-                StartTime = Input.GetTime(),
-                EndTime = Input.GetTime(),
-            };
-            modelRecord.Duration = Input.GetDuration(modelRecord.StartTime, modelRecord.EndTime);
-
-            DbManager.InsertRecord(modelRecord);
-
-            ViewRecords();
-        }
-
-        private static void DisplayDeleteContextMenu()
-        {
-            Console.WriteLine("b to Go Back");
-            Console.WriteLine("d to Delete Record: ");
-        }
-
-        private static void GetRecordsToDelete()
-        {
-            Console.Clear();
-            ViewAllRecords();
 
             DisplayDeleteContextMenu();
-            string choice = Input.GetChoice();
-            while (choice != "b")
-            {
-                switch (choice)
-                {
-                    case "d":
-                        DbManager.DeleteRecord(Input.GetDate());
-                        ViewAllRecords();
-                        break;
-                    default:
-                        Console.WriteLine("Invalid Choice!");
-                        ViewAllRecords();
-                        break;
-                }
+            choice = Input.GetChoice();
+        }
 
-                DisplayDeleteContextMenu();
-                choice = Input.GetChoice();
+        ViewRecords();
+    }
+
+    private static void DisplayUpdateContextMenu()
+    {
+        Console.WriteLine("Choose what to update: ");
+        Console.WriteLine("d to Update Date");
+        Console.WriteLine("t to Update StartTime/EndTime");
+        Console.WriteLine("a to Update All");
+        Console.WriteLine("b to Go Back");
+
+        Console.Write("Your choice? ");
+    }
+
+    private static void SelectRecordToUpdate()
+    {
+        Console.Clear();
+        ViewRecords();
+
+        Console.WriteLine("Select which record to update (by date)");
+        string oldDate = Input.GetDate();
+
+        DisplayUpdateContextMenu();
+        string choice = Input.GetChoice();
+
+        while (choice != "b")
+        {
+            CodingTrackerModel newRecord = new CodingTrackerModel();
+            switch (choice)
+            {
+                case "d":
+                    newRecord.Date = Input.GetDate();
+                    DbManager.UpdateRecord(oldDate, newRecord);
+                    break;
+                case "t":
+                    newRecord.StartTime = Input.GetTime();
+                    newRecord.EndTime = Input.GetTime();
+                    newRecord.Duration = Input.GetDuration(newRecord.StartTime, newRecord.EndTime);
+                    DbManager.UpdateRecord(oldDate, newRecord);
+                    break;
+                case "a":
+                    newRecord.Date = Input.GetDate();
+                    newRecord.StartTime = Input.GetTime();
+                    newRecord.EndTime = Input.GetTime();
+                    newRecord.Duration = Input.GetDuration(newRecord.StartTime, newRecord.EndTime);
+                    DbManager.UpdateRecord(oldDate, newRecord);
+                    break;
+                default:
+                    Console.WriteLine("Invalid choice! Press Enter to continue...");
+                    Console.ReadLine();
+                    break;
             }
 
-            ViewAllRecords();
-        }
-
-        private static void DisplayUpdateContextMenu()
-        {
-            Console.WriteLine("Choose what to update: ");
-            Console.WriteLine("d to Update Date");
-            Console.WriteLine("t to Update StartTime/EndTime");
-            Console.WriteLine("a to Update All");
-            Console.WriteLine("b to Go Back");
-
-            Console.Write("Your choice? ");
-        }
-
-        private static void SelectRecordToUpdate()
-        {
-            Console.Clear();
-            ViewAllRecords();
-
-            DbManager.ReadFromDb();
-
-            Console.WriteLine("Select which record to update (by date)");
-            string oldDate = Input.GetDate();
-
+            ViewRecords();
             DisplayUpdateContextMenu();
-            string choice = Input.GetChoice();
-
-            while (choice != "b")
-            {
-                CodingTrackerModel newRecord = new CodingTrackerModel();
-                switch (choice)
-                {
-                    case "d":
-                        newRecord.Date = Input.GetDate();
-                        DbManager.UpdateRecord(oldDate, newRecord);
-                        break;
-                    case "t":
-                        newRecord.StartTime = Input.GetTime();
-                        newRecord.EndTime = Input.GetTime();
-                        newRecord.Duration = Input.GetDuration(newRecord.StartTime, newRecord.EndTime);
-                        DbManager.UpdateRecord(oldDate, newRecord);
-                        break;
-                    case "a":
-                        newRecord.Date = Input.GetDate();
-                        newRecord.StartTime = Input.GetTime();
-                        newRecord.EndTime = Input.GetTime();
-                        newRecord.Duration = Input.GetDuration(newRecord.StartTime, newRecord.EndTime);
-                        DbManager.UpdateRecord(oldDate, newRecord);
-                        break;
-                    default:
-                        Console.WriteLine("Invalid choice! Press Enter to continue...");
-                        Console.ReadLine();
-                        break;
-                }
-
-                ViewAllRecords();
-                DisplayUpdateContextMenu();
-                choice = Input.GetChoice();
-            }
-
-            Console.Clear();
+            choice = Input.GetChoice();
         }
 
-        private static void ViewRecords()
+        Console.Clear();
+    }
+
+    private static void ViewRecords()
+    {
+        Console.WriteLine("Do you wish to filter");
+        var ans = Input.GetChoice();
+        if (ans == "no")
         {
-            DisplayRecords.View();
+            DisplayRecords.View(CurrentData);
+            return;
         }
 
-        private static void ViewAllRecords()
-        {
-            DisplayRecords.ViewAll();
-        }
+        var filteredList = FilterController.Filter(CurrentData);
+        DisplayRecords.View(filteredList);
     }
 }
