@@ -1,57 +1,98 @@
 ﻿using System.Globalization;
 
-namespace CodingTrackerConsole
+namespace CodingTrackerConsole;
+
+public class UserInput
 {
-    public class UserInput
+    private readonly Validation _dateTimeValidation = new();
+
+    public string GetDate()
     {
-        private readonly Validation _dateTimeValidation = new();
+        Console.Write("Enter date (mm-dd-yyyy): ");
+        var date = Console.ReadLine();
 
-        public string GetDate()
+        while (!_dateTimeValidation.IsValidDate(date))
         {
-            Console.Write("Enter date (mm-dd-yyyy): ");
-            var date = Console.ReadLine();
-
-            while (!_dateTimeValidation.IsValidDate(date))
-            {
-                Console.WriteLine("Invalid date! Expected format (mm-dd-yyyy)");
-                date = Console.ReadLine();
-            }
-
-            return date;
+            Console.WriteLine("Invalid date! Expected format (mm-dd-yyyy)");
+            date = Console.ReadLine();
         }
 
-        public string GetTime()
+        return date;
+    }
+
+    public string GetTime()
+    {
+        Console.Write("Enter time (hh:mm): ");
+        var time = Console.ReadLine();
+
+        while (!_dateTimeValidation.IsValidTime(time))
         {
-            Console.Write("Enter time (hh:mm): ");
-            var time = Console.ReadLine();
-
-            while (!_dateTimeValidation.IsValidTime(time))
-            {
-                Console.Write("Invalid time! Expected format (hh:mm): ");
-                time = Console.ReadLine();
-            }
-
-            return time;
+            Console.Write("Invalid time! Expected format (hh:mm): ");
+            time = Console.ReadLine();
         }
 
-        public string GetDuration(string startTime, string endTime)
+        return time;
+    }
+
+    public string GetDuration(string startTime, string endTime)
+    {
+        DateTime parsedStartTime = DateTime.ParseExact(startTime, "HH:mm", null, DateTimeStyles.None);
+        DateTime parsedEndTime = DateTime.ParseExact(endTime, "HH:mm", null, DateTimeStyles.None);
+
+        TimeSpan duration = parsedEndTime.Subtract(parsedStartTime);
+
+        if (duration < TimeSpan.Zero)
         {
-            DateTime parsedStartTime = DateTime.ParseExact(startTime, "HH:mm", null, DateTimeStyles.None);
-            DateTime parsedEndTime = DateTime.ParseExact(endTime, "HH:mm", null, DateTimeStyles.None);
-
-            TimeSpan duration = parsedEndTime.Subtract(parsedStartTime);
-
-            if (duration < TimeSpan.Zero)
-            {
-                duration += TimeSpan.FromDays(1);
-            }
-
-            return duration.ToString();
+            duration += TimeSpan.FromDays(1);
         }
 
-        public string GetChoice()
+        return duration.ToString();
+    }
+
+    public string GetChoice()
+    {
+        return Console.ReadLine().Trim().ToLower();
+    }
+
+    public string GetDay()
+    {
+        Console.Write("Enter day to filter with (Format: dd):");
+        var day = Console.ReadLine();
+
+        while (!_dateTimeValidation.IsValidDay(day))
         {
-            return Console.ReadLine().Trim().ToLower();
+            Console.WriteLine("Invalid day! Expected format (dd)");
+            day = Console.ReadLine();
         }
+
+        return day;
+    }
+
+    public string GetMonth()
+    {
+        Console.Write("Enter month to filter with (Format: mm): ");
+        var month = Console.ReadLine();
+
+        while (!_dateTimeValidation.IsValidMonth(month))
+        {
+            Console.WriteLine("Invalid month! Expected format (mm): ");
+            month = Console.ReadLine();
+        }
+
+        return month;
+    }
+
+    public string GetYear()
+    {
+        Console.Write("Enter year to filter with (Format: yyyy): ");
+        var year = Console.ReadLine();
+
+        while (!_dateTimeValidation.IsValidYear(year))
+        {
+            Console.WriteLine("Invalid month! Expected format (yyyy)");
+            year = Console.ReadLine();
+        }
+
+        return year;
     }
 }
